# openapi.yaml
openapi: 3.0.3
info:
  title: Smart Chat RAG Chatbot API
  description: |
    RESTful API for Smart Chat - an AI-powered RAG (Retrieval-Augmented Generation) chatbot platform
    with WordPress integration, multi-bot management, and human handoff capabilities.

    ## Features
    - Multi-tenant bot management
    - RAG-based responses using vector embeddings
    - WordPress plugin integration
    - Human agent handoff
    - Stripe billing integration
    - Rate limiting & abuse detection
    - Streaming responses (SSE)

    ## Authentication
    Most endpoints require authentication via NextAuth session cookies or API tokens.
    WordPress endpoints use custom token format: `userId:botId:secret`

  version: 1.0.0
  contact:
    name: API Support
    email: support@smartchat.com
  license:
    name: MIT
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://api.smartchat.com/api
    description: Production server
...
security:
  - sessionAuth: []
  - tokenAuth: []
paths:
...
  /manager/analytics:
    get:
      tags:
        - Analytics
      summary: Get manager analytics
      description: |
        Retrieve various analytics for manager's bots.

        **Available Types:**
        - `languages` - Distribution of user languages
        - `topQuestions` - Most frequently asked questions
      parameters:
        - name: type
          in: query
          required: true
          schema:
            type: string
            enum: [languages, topQuestions]
          description: Type of analytics to retrieve
      responses:
        '200':
          description: Analytics data
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/LanguageAnalytics'
                  - $ref: '#/components/schemas/TopQuestionsAnalytics'
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: NextAuth session cookie
    tokenAuth:
      type: apiKey
      in: header
      name: Authorization
      description: WordPress integration token (userId:botId:secret)
  schemas:
...
    LanguageAnalytics:
      type: object
      properties:
        success:
          type: boolean
          example: true
        type:
          type: string
          example: "languages"
        data:
          type: array
          items:
            type: object
            properties:
              language:
                type: string
                example: "en"
              count:
                type: integer
                example: 450
    TopQuestionsAnalytics:
      type: object
      properties:
        success:
          type: boolean
          example: true
        type:
          type: string
          example: "topQuestions"
        data:
          type: array
          items:
            type: object
            properties:
              question:
                type: string
                example: "What are your business hours?"
              count:
                type: integer
                example: 87